Minor file reorganization in astropy.nddata
