+++ astropy/nddata/__init__.py
@@ -9,13 +9,16 @@ be easily provided by a single array.
 """
 
 from .nddata import *
-from .nddatabase import *
+from .nddata_base import *
 from .nduncertainty import *
 from .flag_collection import *
+
 from .decorators import *
-from .arithmetic import *
-from .slicing import *
-from .io import *
+
+from .mixins.ndarithmetic import *
+from .mixins.ndslicing import *
+from .mixins.ndio import *
+
 from .compat import *
 
 from .. import config as _config
