+++ astropy/nddata/mixins/ndslicing.py
