+++ astropy/nddata/mixins/tests/test_ndarithmetic.py
